Fixed types of function types. Remove redundant code (fields cleanups in Player::LoadFromDB).
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1945 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ObjectAccessor.cpp
@@ -63,7 +63,7 @@ ObjectAccessor::GetCorpse(Unit const &u, uint64 guid)
 }
 
 Corpse*
-ObjectAccessor::GetCorpse(uint32 x, uint32 y, uint32 mapid, uint64 guid)
+ObjectAccessor::GetCorpse(float x, float y, uint32 mapid, uint64 guid)
 {
     CellPair p(MaNGOS::ComputeCellPair(x,y));
     Cell cell = RedZone::GetZone(p);
